Don't recommmend max_split_size_mb first
I've run into a couple of cases now where max_split_size_mb has been set
in projects as a workaround for fragmentation but it ends up causing problems
later, such as degraded performance from freeing empty segments. While it
is a useful setting to have, expandable_segments is probably a better first
resort for fixing fragmentation since when it works it is less likely to
need synchronous GPU operations to continue running.

ghstack-source-id: e7809be6fa76df5d1e3a865c41435fb262faad22
Pull Request resolved: https://github.com/pytorch/pytorch/pull/113481