[aotinductor] add versions for the sdpa shim api
In our first implemenation of the sdpa shim api, we didn't consider
the case where the optional scale argument could be None. It was
unnoticed because we always got a default argument for the cuda backend.
The issue was detected with the cpu backend.

This PR implements versioning for shim kernels. Currently, we only
have different versions for the sdpa api. We expect we would only
maintain a very small number of abi-compatible shim APIs that
had different versions.

ghstack-source-id: 46e533fccd3ca5e4e1da5421c5b561294fbdecc1
Pull Request resolved: https://github.com/pytorch/pytorch/pull/113487