[inductor] Make {freezing,ir}.py pass follow-imports typechecking
I used a couple of type-ignore comments in ir.py because it constructs
short-lived instances of FixedLayout and GraphModuleSerializer, just to
call a single method on them that doesn't use all their members. Making
those unused members optional would make the rest of the code a lot
messier with sprinkled `assert` statements.

ghstack-source-id: 110cfc88042369d56ccc98e357ae469473005229
Pull Request resolved: https://github.com/pytorch/pytorch/pull/113534