Drop dynamo-specific type hints on Tensor in favor of type-ignores
Per [this][1] discussion, plus some offline discussion. The summary:
albanD considers the core PyTorch types like Tensor to be extremely
brittle, and does not think the risk of adding these typed attributes to
be worth it.

eellison mentioned that we could use `WeakTensorKeyDictionary` instead.
However, based on the sparse usage of these bonus attributes, I think
that would be overkill. So I've opted to go with a few more type-ignore
comments instead.

[1]: https://github.com/pytorch/pytorch/pull/113610#discussion_r1392907367

ghstack-source-id: 7389fb187e6b5521ebd48eb5ead6a898264cf494
Pull Request resolved: https://github.com/pytorch/pytorch/pull/113720