[ONNX] Update back the mutated buffers into the original PyTorch model
Currently, when the user_model: torch.export.ExportedProgram is exported
to ONNX and its implementation mutates buffers, the mutated buffer is
returned as an extra output in the ONNX graph while the original model
still have the outdated buffer value.

If we do a numerical comparison between the pytorch model and the ONNX
model, this mismatch is detected, failing the numerical equivalence
check

This PR updates the pytorch model with the new buffer value after the
ONNX model is executed, making both models numerically equivalent.

ghstack-source-id: 27f962d006c6509f40ce1435772c3a1c1f75f45e
Pull Request resolved: https://github.com/pytorch/pytorch/pull/113686