[reland][HigherOrderOp] remove _deprecated_global_ns (#113813)
Summary:
X-link: https://github.com/pytorch/executorch/pull/1224

This is a reland of https://github.com/pytorch/pytorch/issues/112757. Cannot land internally because internal diff is not in sync with OSS due to issues in dealing with two export repos (executorch and pytorch) for the ghimport-ghexport approach.

Will try the web UI of import and export instead of ghimport and ghexport flow.

Pull Request resolved: https://github.com/pytorch/pytorch/pull/113813

Reviewed By: angelayi

Differential Revision: D51373556

Pulled By: ydwu4

fbshipit-source-id: ad3f4d31946636fe8da2fed254bfa75ca7aa3bfd