[AOTI] Improve the two-pass wrapper codegen
Summary: For the second-pass, we don't have to rerun the whole inductor flow again. This PR moves that second-pass to the codegen time. This change not only speeds up the compilation, but also removes kernel scheduling inconsistency between the two passes. Another future improvement is to make the second-pass reuse the scheduler and do the wrapper codegen only.

ghstack-source-id: e29b02957140484c5ed2e30b85eaa7522b6c8cb6
Pull Request resolved: https://github.com/pytorch/pytorch/pull/113762